Did some more work on the install function

+++ include/installer.h
@@ -2,5 +2,6 @@
 #define __INSTALLER__H
 #define ISO_FILE_SIZE_IN_SECTORS 2097152
 #include "types.h"
-int install(int disk_1, int disk_2,uint16_t size);
+int install(int disk_1, int disk_2,uint32_t size);
+
 #endif
